updated about and help sections
